# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sorteia API",
    "summary": "Importações, sorteios e exportações processados apenas em memória.",
    "description": "Nenhum dado de participante é armazenado. Erros seguem a RFC 9457 (`application/problem+json`) com um `code` estável.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": [
          "saúde"
        ],
        "summary": "Verifica se a API está respondendo",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/limits": {
      "get": {
        "tags": [
          "configuração"
        ],
        "summary": "Limites do serviço",
        "operationId": "get_limits_api_v1_limits_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitsOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/text": {
      "post": {
        "tags": [
          "importação"
        ],
        "summary": "Interpreta texto colado ou digitado",
        "description": "Separa as pessoas, normaliza os nomes e aponta vazios, inválidos e duplicados. Também usado para validar um nome digitado ou editado.",
        "operationId": "import_text_api_v1_imports_text_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportTextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Dados ou arquivo inválidos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemOut"
                }
              }
            }
          },
          "429": {
            "description": "Muitas requisições.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/file": {
      "post": {
        "tags": [
          "importação"
        ],
        "summary": "Interpreta uma planilha .xlsx ou um arquivo .csv",
        "description": "O corpo da requisição é o próprio arquivo (não multipart), lido em memória e descartado ao fim da requisição. Opções viajam como índices: nomes de arquivo e de aba nunca aparecem na URL.",
        "operationId": "import_file_api_v1_imports_file_post",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "xlsx",
                    "csv"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Format"
            }
          },
          {
            "name": "sheet",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 500,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sheet"
            }
          },
          {
            "name": "column",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0,
                  "exclusiveMaximum": 50
                },
                {
                  "type": "null"
                }
              ],
              "title": "Column"
            }
          },
          {
            "name": "header",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "auto",
                "yes",
                "no"
              ],
              "type": "string",
              "default": "auto",
              "title": "Header"
            }
          },
          {
            "name": "delimiter",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "auto",
                "semicolon",
                "comma",
                "tab",
                "none"
              ],
              "type": "string",
              "default": "auto",
              "title": "Delimiter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Dados ou arquivo inválidos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemOut"
                }
              }
            }
          },
          "429": {
            "description": "Muitas requisições.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            }
          },
          "413": {
            "description": "Arquivo grande demais.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            }
          },
          "415": {
            "description": "Formato não suportado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        }
      }
    },
    "/api/v1/rounds": {
      "post": {
        "tags": [
          "sorteio"
        ],
        "summary": "Executa uma rodada",
        "description": "Sorteia posições na lista de participantes disponíveis com o gerador criptográfico do sistema operacional. Nenhum nome é enviado e nada é gravado.",
        "operationId": "create_round_api_v1_rounds_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoundRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoundOut"
                }
              }
            }
          },
          "422": {
            "description": "Pedido inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemOut"
                }
              }
            }
          },
          "429": {
            "description": "Muitas requisições.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exports": {
      "post": {
        "tags": [
          "exportação"
        ],
        "summary": "Gera o arquivo de resultado (CSV ou XLSX)",
        "description": "Recebe as rodadas escolhidas e devolve o arquivo. Nada é gravado no servidor.",
        "operationId": "create_export_api_v1_exports_post",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "csv",
                "xlsx"
              ],
              "type": "string",
              "default": "xlsx",
              "title": "Format"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Arquivo gerado.",
            "content": {
              "text/csv": {},
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {}
            }
          },
          "422": {
            "description": "Dados inválidos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemOut"
                }
              }
            }
          },
          "429": {
            "description": "Muitas requisições.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Algorithm": {
        "type": "string",
        "enum": [
          "partial-fisher-yates/1+os-csprng",
          "uniform-with-replacement/1+os-csprng"
        ]
      },
      "ColumnOut": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "letter": {
            "type": "string",
            "title": "Letter"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "samples": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Samples"
          },
          "filled": {
            "type": "integer",
            "title": "Filled"
          }
        },
        "type": "object",
        "required": [
          "index",
          "letter",
          "label",
          "samples",
          "filled"
        ],
        "title": "ColumnOut"
      },
      "DuplicateGroupOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "rows": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Rows",
            "description": "Primeiras linhas do grupo (até 20)."
          }
        },
        "type": "object",
        "required": [
          "name",
          "count",
          "rows"
        ],
        "title": "DuplicateGroupOut"
      },
      "EntryOut": {
        "properties": {
          "row": {
            "type": "integer",
            "title": "Row",
            "description": "Linha (ou item) de origem, a partir de 1."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Nome normalizado, pronto para exibir."
          },
          "key": {
            "type": "string",
            "title": "Key",
            "description": "Chave de duplicidade (sem acentos, maiúsculas e espaços extras)."
          },
          "repeat_of": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repeat Of",
            "description": "Linha da primeira ocorrência, se for repetição."
          }
        },
        "type": "object",
        "required": [
          "row",
          "name",
          "key",
          "repeat_of"
        ],
        "title": "EntryOut"
      },
      "ExportRequest": {
        "properties": {
          "draw_name": {
            "type": "string",
            "maxLength": 500,
            "title": "Draw Name"
          },
          "timezone": {
            "type": "string",
            "maxLength": 64,
            "title": "Timezone",
            "description": "Fuso IANA, ex.: America/Sao_Paulo."
          },
          "rounds": {
            "items": {
              "$ref": "#/components/schemas/ExportRoundIn"
            },
            "type": "array",
            "minItems": 1,
            "title": "Rounds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "draw_name",
          "timezone",
          "rounds"
        ],
        "title": "ExportRequest",
        "description": "Dados que o navegador envia para gerar o arquivo. Nada é guardado no servidor."
      },
      "ExportRoundIn": {
        "properties": {
          "number": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Number"
          },
          "drawn_at": {
            "type": "string",
            "format": "date-time",
            "title": "Drawn At"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Quantity"
          },
          "allow_repeat": {
            "type": "boolean",
            "title": "Allow Repeat"
          },
          "remove_winners": {
            "type": "boolean",
            "title": "Remove Winners"
          },
          "total_participants": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Participants"
          },
          "pool_size": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pool Size"
          },
          "algorithm": {
            "$ref": "#/components/schemas/Algorithm"
          },
          "winners": {
            "items": {
              "$ref": "#/components/schemas/ExportWinnerIn"
            },
            "type": "array",
            "maxItems": 50000,
            "minItems": 1,
            "title": "Winners"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "number",
          "drawn_at",
          "quantity",
          "allow_repeat",
          "remove_winners",
          "total_participants",
          "pool_size",
          "algorithm",
          "winners"
        ],
        "title": "ExportRoundIn"
      },
      "ExportWinnerIn": {
        "properties": {
          "position": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Position"
          },
          "name": {
            "type": "string",
            "maxLength": 500,
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "position",
          "name"
        ],
        "title": "ExportWinnerIn"
      },
      "FieldErrorOut": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "field",
          "type"
        ],
        "title": "FieldErrorOut"
      },
      "HealthOut": {
        "properties": {
          "status": {
            "type": "string",
            "const": "ok",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "HealthOut"
      },
      "ImportPreviewOut": {
        "properties": {
          "source": {
            "type": "string",
            "enum": [
              "text",
              "csv",
              "xlsx"
            ],
            "title": "Source"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/EntryOut"
            },
            "type": "array",
            "title": "Entries"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/IssueOut"
            },
            "type": "array",
            "title": "Issues",
            "description": "Até 100 linhas com problema; total em `stats`."
          },
          "duplicate_groups": {
            "items": {
              "$ref": "#/components/schemas/DuplicateGroupOut"
            },
            "type": "array",
            "title": "Duplicate Groups",
            "description": "Até 100 grupos."
          },
          "stats": {
            "$ref": "#/components/schemas/StatsOut"
          },
          "separator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Separator",
            "description": "Separador usado (texto) ou delimitador (CSV)."
          },
          "has_header": {
            "type": "boolean",
            "title": "Has Header"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnOut"
            },
            "type": "array",
            "title": "Columns"
          },
          "column": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Column"
          },
          "sheets": {
            "items": {
              "$ref": "#/components/schemas/SheetOut"
            },
            "type": "array",
            "title": "Sheets"
          },
          "sheet": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sheet"
          },
          "max_name_length": {
            "type": "integer",
            "title": "Max Name Length"
          }
        },
        "type": "object",
        "required": [
          "source",
          "entries",
          "issues",
          "duplicate_groups",
          "stats",
          "separator",
          "has_header",
          "columns",
          "column",
          "sheets",
          "sheet",
          "max_name_length"
        ],
        "title": "ImportPreviewOut"
      },
      "ImportTextRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 4000000,
            "title": "Text",
            "description": "Texto colado ou digitado."
          },
          "separator": {
            "type": "string",
            "enum": [
              "auto",
              "newline",
              "semicolon",
              "comma"
            ],
            "title": "Separator",
            "description": "Como separar as pessoas. Em `auto`, mais de uma linha → uma por linha; senão ponto e vírgula; senão vírgula. Colunas com tab viram tabela.",
            "default": "auto"
          },
          "header": {
            "type": "string",
            "enum": [
              "auto",
              "yes",
              "no"
            ],
            "title": "Header",
            "default": "auto"
          },
          "column": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMaximum": 50.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Column"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ImportTextRequest"
      },
      "IssueOut": {
        "properties": {
          "row": {
            "type": "integer",
            "title": "Row"
          },
          "code": {
            "type": "string",
            "enum": [
              "too_long",
              "cell_error"
            ],
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "row",
          "code"
        ],
        "title": "IssueOut"
      },
      "LimitsOut": {
        "properties": {
          "max_participants": {
            "type": "integer",
            "title": "Max Participants"
          },
          "max_round_quantity": {
            "type": "integer",
            "title": "Max Round Quantity"
          },
          "max_name_length": {
            "type": "integer",
            "title": "Max Name Length"
          },
          "max_draw_name_length": {
            "type": "integer",
            "title": "Max Draw Name Length"
          },
          "max_upload_bytes": {
            "type": "integer",
            "title": "Max Upload Bytes"
          },
          "max_text_chars": {
            "type": "integer",
            "title": "Max Text Chars"
          },
          "max_export_rounds": {
            "type": "integer",
            "title": "Max Export Rounds"
          }
        },
        "type": "object",
        "required": [
          "max_participants",
          "max_round_quantity",
          "max_name_length",
          "max_draw_name_length",
          "max_upload_bytes",
          "max_text_chars",
          "max_export_rounds"
        ],
        "title": "LimitsOut",
        "description": "Limites do serviço, para a interface validar antes de enviar."
      },
      "ProblemOut": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "params": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "string"
                }
              ]
            },
            "type": "object",
            "title": "Params",
            "default": {}
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "ProblemOut",
        "description": "Erro no formato RFC 9457 (``application/problem+json``)."
      },
      "RoundOut": {
        "properties": {
          "positions": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Positions",
            "description": "Posições sorteadas (a partir de 0) na lista de disponíveis, na ordem do sorteio: a primeira é o 1º vencedor."
          },
          "pool_size": {
            "type": "integer",
            "title": "Pool Size"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "allow_repeat": {
            "type": "boolean",
            "title": "Allow Repeat"
          },
          "algorithm": {
            "type": "string",
            "title": "Algorithm",
            "description": "Algoritmo e fonte de aleatoriedade usados."
          },
          "drawn_at": {
            "type": "string",
            "format": "date-time",
            "title": "Drawn At",
            "description": "Horário oficial da rodada (relógio do servidor)."
          }
        },
        "type": "object",
        "required": [
          "positions",
          "pool_size",
          "quantity",
          "allow_repeat",
          "algorithm",
          "drawn_at"
        ],
        "title": "RoundOut"
      },
      "RoundRequest": {
        "properties": {
          "pool_size": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pool Size",
            "description": "Quantos participantes estão disponíveis."
          },
          "quantity": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Quantity",
            "description": "Quantos vencedores sortear."
          },
          "allow_repeat": {
            "type": "boolean",
            "title": "Allow Repeat",
            "description": "Se a mesma pessoa pode sair mais de uma vez."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pool_size",
          "quantity",
          "allow_repeat"
        ],
        "title": "RoundRequest",
        "description": "Pedido de rodada. Nenhum nome é enviado: o servidor escolhe posições na lista de\nparticipantes disponíveis que o navegador congelou no momento do clique (ADR-018)."
      },
      "SheetOut": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "hidden": {
            "type": "boolean",
            "title": "Hidden"
          }
        },
        "type": "object",
        "required": [
          "index",
          "name",
          "hidden"
        ],
        "title": "SheetOut"
      },
      "StatsOut": {
        "properties": {
          "rows": {
            "type": "integer",
            "title": "Rows",
            "description": "Linhas de dados lidas (sem o cabeçalho)."
          },
          "valid": {
            "type": "integer",
            "title": "Valid"
          },
          "empty": {
            "type": "integer",
            "title": "Empty"
          },
          "invalid": {
            "type": "integer",
            "title": "Invalid"
          },
          "duplicates": {
            "type": "integer",
            "title": "Duplicates",
            "description": "Ocorrências repetidas além da primeira."
          },
          "duplicate_groups": {
            "type": "integer",
            "title": "Duplicate Groups"
          },
          "dates": {
            "type": "integer",
            "title": "Dates",
            "description": "Valores que eram datas: indica coluna provavelmente errada."
          }
        },
        "type": "object",
        "required": [
          "rows",
          "valid",
          "empty",
          "invalid",
          "duplicates",
          "duplicate_groups",
          "dates"
        ],
        "title": "StatsOut"
      },
      "ValidationProblemOut": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "params": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "string"
                }
              ]
            },
            "type": "object",
            "title": "Params",
            "default": {}
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldErrorOut"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "ValidationProblemOut"
      }
    }
  }
}